+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 ## TBD 
 * Added cross-cloud B2B support.
 * Fixed logic to handle links that open in new tab for embedded webview.
+* accountForUsername from MSALPublicClientApplication will return nil back when username is nil or empty
 * Updated user guide to provide a sample Swift & ObjC code for querying a specific account and return token silently when multiple accounts are present in the cache. 
+* Added client-side fix for the known ADFS PKeyAuth issue. (#1150)
 
 ## [1.1.11] - 2020-10-16
 * Enabled PKeyAuth via UserAgent String on MacOS